restructured transformer submodule
