+++ sktime/transformers/dictionary_based/PAA.py
