+++ sktime/transformers/tests/test_Deseasonaliser.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from sktime.transformers.series_to_series import Deseasonaliser
+from sktime.transformers.forecasting import Deseasonaliser
 from sktime.utils.testing import generate_seasonal_time_series_data_with_trend, generate_time_series_data_with_trend
 from sktime.utils.data_container import select_times, tabularise
