+++ sktime/transformers/tests/test_RandomIntervalSegmenter.py
@@ -1,4 +1,4 @@
-from sktime.transformers.series_to_tabular import RandomIntervalSegmenter
+from sktime.transformers.segment import RandomIntervalSegmenter
 from sktime.utils.testing import generate_df_from_array
 from sktime.utils.data_container import tabularize
 import pytest
